+++ test/cli.test.js
@@ -1713,10 +1713,10 @@ describe('[Mainchain]', () => {
     expect(initializeAndClaimArrayTx.gasPrice.toString()).toBe(chainGasPrice);
   });
 
-  // sendETH
-  test('[mainchain] iexec wallet sendETH', async () => {
+  // send-ether
+  test('[mainchain] iexec wallet send-ether', async () => {
     const raw = await execAsync(
-      `${iexecPath} wallet sendETH 1 --to ${POOR_ADDRESS1} --force --raw`,
+      `${iexecPath} wallet send-ether 1 --to ${POOR_ADDRESS1} --force --raw`,
     );
     const res = JSON.parse(raw);
     expect(res.ok).toBe(true);
@@ -1729,10 +1729,9 @@ describe('[Mainchain]', () => {
     expect(tx.gasPrice.toString()).toBe(chainGasPrice);
   });
 
-  // sendRLC
-  test('[mainchain] iexec wallet sendRLC 1000000000', async () => {
+  test('[mainchain] iexec wallet send-ether', async () => {
     const raw = await execAsync(
-      `${iexecPath} wallet sendRLC 1000000000 --to ${POOR_ADDRESS1} --force --raw`,
+      `${iexecPath} wallet send-ether 1 gwei --to ${POOR_ADDRESS1} --force --raw`,
     );
     const res = JSON.parse(raw);
     expect(res.ok).toBe(true);
@@ -1745,15 +1744,30 @@ describe('[Mainchain]', () => {
     expect(tx.gasPrice.toString()).toBe(chainGasPrice);
   });
 
-  test('[mainchain] iexec wallet sendRLC 0.5 RLC', async () => {
+  // send-RLC
+  test('[mainchain] iexec wallet send-RLC 0.5', async () => {
     const raw = await execAsync(
-      `${iexecPath} wallet sendRLC 0.5 RLC --to ${POOR_ADDRESS1} --force --raw`,
+      `${iexecPath} wallet send-RLC 0.5 --to ${POOR_ADDRESS1} --force --raw`,
     );
     const res = JSON.parse(raw);
     expect(res.ok).toBe(true);
     expect(res.from).toBe(ADDRESS);
     expect(res.to).toBe(POOR_ADDRESS1);
     expect(res.amount).toBe('500000000');
+    expect(res.txHash).toBeDefined();
+    const tx = await tokenChainRPC.getTransaction(res.txHash);
+    expect(tx).toBeDefined();
+    expect(tx.gasPrice.toString()).toBe(chainGasPrice);
+  });
+
+  test('[mainchain] iexec wallet send-RLC 1000000000 nRLC', async () => {
+    const raw = await execAsync(
+      `${iexecPath} wallet send-RLC 1000000000 nRLC --to ${POOR_ADDRESS1} --force --raw`,
+    );
+    const res = JSON.parse(raw);
+    expect(res.ok).toBe(true);
+    expect(res.from).toBe(ADDRESS);
+    expect(res.to).toBe(POOR_ADDRESS1);
     expect(res.txHash).toBeDefined();
     const tx = await tokenChainRPC.getTransaction(res.txHash);
     expect(tx).toBeDefined();
@@ -2181,9 +2195,9 @@ describe('[Sidechain]', () => {
     expect(res.balance.ether).toBeUndefined();
   });
 
-  test('[sidechain] iexec wallet sendETH', async () => {
+  test('[sidechain] iexec wallet send-ether', async () => {
     const raw = await execAsync(
-      `${iexecPath} wallet sendETH 0.1 --to ${POOR_ADDRESS1} --force --raw`,
+      `${iexecPath} wallet send-ether 0.1 --to ${POOR_ADDRESS1} --force --raw`,
     ).catch((e) => e.message);
     const res = JSON.parse(raw);
     expect(res.ok).toBe(false);